CAR ACCIDENT DETECTION USING MACHINE LEARNING
We are witnessing a growing number of deaths due to road accidents. One of the main 
reasons being the late arrival of help. So, on this project we focus on a way through which we 
can help in detecting the accident and inform the concerned authorities as quickly as possible. 
The detection is done through image recognition system which will be installed in the 
surveillance cameras. And as soon as accident occurs the GPS coordinates will be sent to the 
nearby hospitals and police stations. We believe this will be of great help for saving a person 
in the golden hours of an accident.